pauffinage ml00

+++ ml00/ex03/tools.py
@@ -3,13 +3,13 @@ import numpy as np
 def add_intercept(x):
     """Adds a column of 1's to the non-empty numpy.array x.
     Args:
-    x: has to be a numpy.array of dimension m * n.
+        x: has to be a numpy.array of dimension m * n.
     Returns:
-    X, a numpy.array of dimension m * (n + 1).
-    None if x is not a numpy.array.
-    None if x is an empty numpy.array.
+        X, a numpy.array of dimension m * (n + 1).
+        None if x is not a numpy.array.
+        None if x is an empty numpy.array.
     Raises:
-    This function should not raise any Exception.
+        This function should not raise any Exception.
     """
     if not (isinstance(x, np.ndarray)):
         return None
@@ -20,7 +20,7 @@ def add_intercept(x):
 
     a = np.ones(x.shape[0], dtype=x.dtype).reshape(x.shape[0], 1)
     res = np.concatenate((a, x), axis=1)
-    return (res)
+    return (res.astype(float))
 
 if __name__ == "__main__":
     x = np.arange(1, 6)
